Adding LINEAR11 parser for the case when the exponent is known, Fixing compile time warnings

+++ src/PMB.cpp
@@ -8,7 +8,7 @@ uint8_t PMB_read8BitRegister(uint8_t deviceAddress, uint8_t registerAddress) {
   Wire.beginTransmission(deviceAddress);
   Wire.write(registerAddress);
   Wire.endTransmission(false);
-  Wire.requestFrom(deviceAddress, (uint8_t)1, false);
+  Wire.requestFrom(deviceAddress, (uint8_t)1, (uint8_t)false);
 
   uint8_t rawValue = 0;
 
@@ -25,7 +25,7 @@ uint16_t PMB_read16BitRegister(uint8_t deviceAddress, uint8_t registerAddress) {
   Wire.beginTransmission(deviceAddress);
   Wire.write(registerAddress);
   Wire.endTransmission(false);
-  Wire.requestFrom(deviceAddress, (uint8_t)2, false);
+  Wire.requestFrom(deviceAddress, (uint8_t)2, (uint8_t)false);
 
   uint16_t rawValue = 0;
